// common/enums/InputType.ts
export enum InputType {
    FLOAT_ARRAY_INPUT = 'FLOAT_ARRAY_INPUT',
    INT_ARRAY_INPUT = 'INT_ARRAY_INPUT',
    TEXT_ARRAY_INPUT = 'TEXT_ARRAY_INPUT',
    TEXTAREA_ARRAY_INPUT = 'TEXTAREA_ARRAY_INPUT',
    BOOLEAN_INPUT = 'BOOLEAN_INPUT',
    CHECKBOX = 'CHECKBOX',
    RADIO = 'RADIO',
    SELECT = 'SELECT',
    COUNTRY_CODE_SELECT = 'COUNTRY_CODE_SELECT',
    ENVIRONMENT_SELECT = 'ENVIRONMENT_SELECT',
    REVISION_SELECT = 'REVISION_SELECT',
    GENERIC_CONDITION = 'GENERIC_CONDITION',
    STRING_CONDITION = 'STRING_CONDITION',
    NUMBER_CONDITION = 'NUMBER_CONDITION',
    DOM_SELECTOR_INPUT = 'DOM_SELECTOR_INPUT',
    DATE_STRING = 'DATE_STRING',
    DATETIME_STRING = 'DATETIME_STRING',
    FLOAT_INPUT = 'FLOAT_INPUT',
    INT_INPUT = 'INT_INPUT',
    OBJECT_INPUT = 'OBJECT_INPUT',
    OBJECT_ARRAY_INPUT = 'OBJECT_ARRAY_INPUT',
    TEXT = 'TEXT',
    TEXT_WITH_MACRO_SUPPORT = 'TEXT_WITH_MACRO_SUPPORT',
    TEXTAREA = 'TEXTAREA',
    JAVASCRIPT = 'JAVASCRIPT',
    JSON = 'JSON',
    HTML = 'HTML',
    CSS = 'CSS',
    COLOR = 'COLOR',
    URL = 'URL',
    URL_WITH_MACRO_SUPPORT = 'URL_WITH_MACRO_SUPPORT',
    EMAIL = 'EMAIL',
    DATE_STAMP = 'DATE_STAMP',
    DATETIME_STAMP = 'DATETIME_STAMP',
    INGEST_ENDPOINT_PAYLOAD_DESIGNER = 'INGEST_ENDPOINT_PAYLOAD_DESIGNER',
    CONSENT_VENDORS = 'CONSENT_VENDORS',
    CONSENT_PURPOSES = 'CONSENT_PURPOSES',
    MULTIPLE_SELECT = 'MULTIPLE_SELECT',
}